+++ data.tf
@@ -1,15 +1,13 @@
+# Disabled CloudInit APT update and upgrade to avoid error
+# "Failed to update package using apt: Unexpected error while running command. Command: ['eatmydata', 'apt-get', '--option=Dpkg::Options::=--force-confold', '--option=Dpkg::options::=--force-unsafe-io', '--assume-yes', '--quiet', 'update'] Exit code: 100 Reason: - Stdout: - Stderr: -"
 locals {
   remote_provisioner_host = var.droplet_provisioner_external_ip ? digitalocean_droplet.this.ipv4_address : digitalocean_droplet.this.ipv4_address_private
   default_environment_variables = [
     "DROPLET_INTERNAL_IP=${digitalocean_droplet.this.ipv4_address_private}",
     "DROPLET_EXTERNAL_IP=${digitalocean_droplet.this.ipv4_address}",
   ]
-  default_packages = [
-    "curl",
-    "mc",
-    "net-tools"
-  ]
   default_commands = [
+    "sudo DEBIAN_FRONTEND=noninteractive apt-get update && sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade -y",
     "sudo mkdir -p ${var.app_data}/${var.app_configurations}",
     "sudo chown ${var.droplet_user}:terraform ${var.app_data}/${var.app_configurations}",
     "sudo chmod 775 ${var.app_data}/${var.app_configurations}",
@@ -19,8 +17,8 @@ locals {
 
 ssh_pwauth: false
 disable_root: true
-package_update: true
-package_upgrade: true
+package_update: false
+package_upgrade: false
 manage_etc_hosts: true
 
 users:
@@ -29,19 +27,17 @@ users:
       - sudo
     sudo:
       - ALL=(ALL) NOPASSWD:ALL
-    ssh-authorized-keys:
+    ssh_authorized_keys:
       - ${data.digitalocean_ssh_key.user.public_key}
   - name: terraform
     groups:
       - sudo
     sudo:
       - ALL=(ALL) NOPASSWD:ALL
-    ssh-authorized-keys:
+    ssh_authorized_keys:
       - ${data.digitalocean_ssh_key.remote_provisioner.public_key}
 
-packages:
-${local.default_packages != null ? join("\n", formatlist("  - '%s'", local.default_packages)) : ""}
-${var.os_packages != null ? join("\n", formatlist("  - '%s'", var.os_packages)) : ""}
+${var.os_packages != null && length(var.os_packages) > 0 ? "packages:\n" : ""}${var.os_packages != null ? join("\n", formatlist("  - '%s'", var.os_packages)) : ""}
 
 runcmd:
 ${local.default_commands != null ? join("\n", formatlist("  - '%s'", local.default_commands)) : ""}
@@ -61,7 +57,8 @@ data "digitalocean_project" "this" {
 }
 
 data "digitalocean_domain" "this" {
-  name = var.droplet_dns_zone
+  count = var.dns_provider == "digitalocean" ? 1 : 0
+  name  = var.droplet_dns_zone
 }
 
 data "digitalocean_vpc" "this" {
@@ -72,4 +69,14 @@ data "digitalocean_droplet_snapshot" "this" {
   name        = var.droplet_image
   region      = var.droplet_region
   most_recent = true
+}
+
+data "cloudflare_zones" "this" {
+  count = var.droplet_dns_record && var.dns_provider == "cloudflare" ? 1 : 0
+  name  = var.droplet_dns_zone
+}
+
+data "cloudflare_zone" "this" {
+  count   = length(data.cloudflare_zones.this)
+  zone_id = length(data.cloudflare_zones.this[count.index].result) > 0 ? data.cloudflare_zones.this[count.index].result[0].id : null
 }