+++ versions.tf
@@ -1,10 +1,14 @@
 terraform {
-  required_version = ">= 1.9"
+  required_version = ">= 1.11"
 
   required_providers {
     digitalocean = {
       source  = "digitalocean/digitalocean"
       version = ">= 2"
+    }
+    cloudflare = {
+      source  = "cloudflare/cloudflare"
+      version = ">= 5"
     }
   }
 }